requirements.txt
tweepy
pyyaml
python-dotenv
openai
requests
